refactor(core-weave): extract mesh inventory renderers
- move mesh-inventory Turtle renderers into mesh_inventory_renderers
- keep weave.ts as the planner dispatcher and public facade
- preserve fixture-sensitive legacy renderer output
- defer shared LocatedFile/ResourcePage block helper cleanup
